(ScopDetection): Track all errors during detection.
This adds a simple log to all errors encountered during
SCoPDetection. If we ignore situations like non-linearity in
memory access functions, we still continue detection and
track the error for further use.

This can be used to decide which information needs to be added
to a SCoP, in order to comply with the polyhedron model.

Conflicts:
	include/polly/ScopDetection.h
	lib/Analysis/ScopDetection.cpp